requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
pytest
selenium>=4.10.0
webdriver-manager>=3.8.6
